video-automator/requirements.txt
PyQt5>=5.15.0
openai-whisper
torch
torchvision
torchaudio
pillow